% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances_calculation.R
\name{get_distance}
\alias{get_distance}
\title{Distances (and their heights) between fuel layers}
\usage{
get_distance (gaps_fbhs_metrics)
}
\arguments{
\item{gaps_fbhs_metrics}{data frame with gaps (distances) and fuel base heights (output of [get_gaps_fbhs()] function).
An object of the class text}
}
\value{
A data frame giving distances (and their heights) between fuel layers in meters.
}
\description{
This function calculates distances between fuel layers as the difference between consecutive gaps and fuel bases
(the gap height always must be lower than the fuel base height).
}
\details{
# List of tree metrics:
\itemize{
\item treeID: tree ID with strings and numeric values
\item treeID1: tree ID with only numeric values
\item cbh - Height of the fuel layers base height (m)
\item gap - Height of gaps between consecutive fuel layers (m)
\item dist: Distance between consecutive fuel layers (m)
\item Hdist - Height of the distance between consecutive fuel layers (m)
\item max_height - Maximum height of the tree profile
}
}
\examples{
## Not run:
library(dplyr)
library(magrittr)
library(gdata)

# Tree metrics derived from get_gaps_fbhs() function
gaps_fbhs_metrics_path <- system.file("extdata", "1_gaps_fbhs_metrics.txt", package = "LadderFuelsR")
gaps_fbhs_metrics <- read.table(gaps_fbhs_metrics_path, sep="\t", header=TRUE)

gaps_fbhs_metrics$treeID <- factor(gaps_fbhs_metrics$treeID)
trees_name1 <- as.character(gaps_fbhs_metrics$treeID)
trees_name2 <- factor(unique(trees_name1))

metrics_distance_list <- list()

for (i in levels(trees_name2)) {

  # Filter data for each tree
  tree2 <- gaps_fbhs_metrics |> dplyr::filter(treeID == i)

  # Get distance metrics for each tree
  metrics_distance <- get_distance(tree2)
  metrics_distance_list[[i]] <- metrics_distance
}

# Combine the individual data frames
metrics_all_distance <- dplyr::bind_rows(metrics_distance_list)

distance_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "2_distance_metrics.txt")
write.table(metrics_all_distance, file=distance_path, sep="\t", row.names=FALSE)
## End(Not run)
}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
