% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maxlad_metrics.R
\name{get_plots_cbh_LAD}
\alias{get_plots_cbh_LAD}
\title{Plots of fuel layers with LAD percentage > 5 and the canopy base height (CBH) based on the maximum LAD percentage}
\usage{
get_plots_cbh_LAD(LAD_profiles, effective_LAD)
}
\arguments{
\item{LAD_profiles}{Original tree Leaf Area Density (LAD) profile (output of [lad.profile()] function in the \emph{leafR} package).
An object of the class text.}

\item{effective_LAD}{Tree metrics with gaps (distances), fuel base heights, and depths of fuel layers with LAD percentage greater than 5
(output of [get_layers_lad()] function).
An object of the class text.}
}
\value{
A plot drawing the CBH of the fuel layer with the maximum LAD percentage and other fuel layers with LAD percentage greater than 5.
}
\description{
This function plots the CBH of the fuel layer with the maximum LAD percentage and other fuel layers with LAD percentage greater than 5.
}
\examples{
\dontrun{
library(ggplot2)
library(dplyr)

# LAD profiles derived from normalized ALS data after applying [lad.profile()] function
LAD_profiles <- read.table(system.file("extdata", "LAD_profiles.txt", package = "LadderFuelsR"),
header = TRUE)
LAD_profiles$treeID <- factor(LAD_profiles$treeID)

# Load the effective_LAD object
if (interactive()) {
effective_LAD <- get_layers_lad()
LadderFuelsR::effective_LAD$treeID <- factor(LadderFuelsR::effective_LAD$treeID)

trees_name1 <- as.character(effective_LAD$treeID)
trees_name2 <- factor(unique(trees_name1))

# Generate plots for fuels LAD metrics
plots_trees_LAD <- get_plots_cbh_LAD(LAD_profiles, effective_LAD)
}
}
}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
