% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumm_breaks_metrics.R
\name{get_plots_cumm}
\alias{get_plots_cumm}
\title{Plot of tree profiles with the canopy base height (CBH) based on the breaking point method}
\usage{
get_plots_cumm(LAD_profiles, cummulative_LAD)
}
\arguments{
\item{LAD_profiles}{original tree Leaf Area Index (LAD) profile (output of [lad.profile()] function from leafR package).
An object of the class text}

\item{cummulative_LAD}{tree metrics derived from using breaking points on cummulative LAD (output of [get_cum_break()] function).
An object of the class text}
}
\value{
A plot of the CBH and LAD percentage below and above the CBH
}
\description{
This function plots the canopy base height (CBH) based on breaking point over the cummulative LAD values and gives the LAD percentage below and above that breaking point
}
\examples{
## Not run:
library(ggplot2)
# LAD profiles derived from normalized ALS data after applying [lad.profile()] function
LAD_profiles$treeID <- factor(LAD_profiles$treeID)

## Not run:
# Load or create the cummulative_LAD object
if (interactive()) {
  cummulative_LAD <- get_cum_break()
  LadderFuelsR::cummulative_LAD$treeID <- factor(LadderFuelsR::cummulative_LAD$treeID)

# Tree metrics derived from get_cum_break() function
cummulative_LAD$treeID <- factor(cummulative_LAD$treeID)

# Generate cumulative LAD plots
plots_trees_cumlad <- get_plots_cumm(LAD_profiles, cummulative_LAD)
}
## End(Not run)

}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
