% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gaps_fbhs.R
\name{get_plots_gap_fbh}
\alias{get_plots_gap_fbh}
\title{Plots of tree profiles with gaps and fuel layers base height (fbh)}
\usage{
get_plots_gap_fbh (LAD_profiles,depth_metrics)
}
\arguments{
\item{LAD_profiles}{original tree Leaf Area Index (LAD) profile (output of [lad.profile()] function in the \emph{leafR} package.
An object of the class text}

\item{depth_metrics}{tree metrics with gaps (distances), fuel base heights and depths (output of [get_depths()] function).
An object of the class text}
}
\value{
A plot drawing by lines the height of gaps and fuel layers bases in tiff format.
}
\description{
This function plots gaps and fuel layers base height (fbh) in the vertical tree profile (VTP).
}
\examples{
## Not run:
library(ggplot2)

# LAD profiles derived from normalized ALS data after applying [lad.profile()] function
data_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "LAD_profiles.txt")
LAD_profiles <- read.table(data_path, sep = "\t", header = TRUE)
LAD_profiles$treeID <- factor(LAD_profiles$treeID)

# Tree metrics derived from get_depths() function
depth_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "3_depth_metrics.txt")
depth_metrics <- read.table(depth_path, sep = "\t", header = TRUE)
depth_metrics$treeID <- factor(depth_metrics$treeID)

# Generate plots for gaps and fbhs
plots_gaps_fbhs <- get_plots_gap_fbh(LAD_profiles, depth_metrics)

# Save plots for each tree
for (name in names(plots_gaps_fbhs)) {
  print(plots_gaps_fbhs[[name]])  # print the plot
  ggsave(file.path(system.file("extdata", package = "LadderFuelsR"),  paste0( name, "_gap_fbh", ".tiff")), plot = plots_gaps_fbhs[[name]])
}
## End(Not run)

}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
