% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_base_heights.R
\name{get_real_fbh}
\alias{get_real_fbh}
\title{Fuels base height after removing distances = 1 m}
\usage{
get_real_fbh(depth_metrics)
}
\arguments{
\item{depth_metrics}{Tree metrics with gaps (distances), fuel base heights, and depths
(output of [get_depths()] function). An object of the class text.}
}
\value{
A data frame giving the first "base height" from those consecutive ones separated by a distance = 1.
This value replaces the values of the next base heights if they are separated by a distance = 1.
}
\description{
This function recalculates fuels base height after removing distances = 1 m,
keeping the first "base height" from those consecutive ones separated by a distance = 1.
}
\details{
# List of tree metrics:
\itemize{
  \item treeID: tree ID with strings and numeric values
  \item treeID1: tree ID with only numeric values
  \item dist: Distance between consecutive fuel layers (m)
  \item Hdist - Height of the distance between consecutive fuel layers (m)
  \item Hcbh - Height of the first base height from those ones separated by a distance = 1.
  \item depth - Depth of fuel layers (m)
  \item Hdepth - Height of the depth of fuel layers (m)
  \item max_height - Maximum height of the tree profile
}
}
\examples{
## Not run:
library(SSBtools)
library(dplyr)
library(magrittr)

# Tree metrics derived from get_depths() function
depth_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "3_depth_metrics.txt")
depth_metrics <- read.table(depth_path, sep = "\t", header = TRUE)
depth_metrics$treeID <- factor(depth_metrics$treeID)

trees_name1 <- as.character(depth_metrics$treeID)
trees_name2 <- factor(unique(trees_name1))

fbh_corr_list <- list()

for (i in levels(trees_name2)){

  # Filter data for each tree
  tree3 <- depth_metrics |> dplyr::filter(treeID == i)

  # Get real fbh for each tree
  fbh_corr <- get_real_fbh(tree3)

  # Store fbh values in a list
  fbh_corr_list[[i]] <- fbh_corr
}

# Combine fbh values for all trees
fbh_corr_all <- dplyr::bind_rows(fbh_corr_list)
fbh_corr_all$treeID <- factor(fbh_corr_all$treeID)

# Reorder columns
# Get original column names
original_column_names <- colnames(fbh_corr_all)

# Specify prefixes
prefixes <- c("treeID", "Hdist", "Hcbh", "Hdepth", "dist", "depth", "max_height")

# Initialize vector to store new order
new_order <- c()

# Loop over prefixes
for (prefix in prefixes) {
  # Find column names matching the current prefix
  matching_columns <- grep(paste0("^", prefix), original_column_names, value = TRUE)
  # Append to the new order
  new_order <- c(new_order, matching_columns)
}

# Reorder values
fbh_corr_all <- fbh_corr_all[, new_order]

# Save the reordered data
fbh_corr_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "4_fbh_metrics_corr.txt")
write.table(fbh_corr_all, file = fbh_corr_path, sep = "\t", row.names = FALSE)

## End(Not run)

}
\author{
Olga Viedma.
}
