% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed0_df.R
\name{get_renamed0_df}
\alias{get_renamed0_df}
\title{Rename and reorder columns (I)}
\usage{
get_renamed0_df (df)
}
\arguments{
\item{df}{internal data frame derived from [get_real_depths()] function}
}
\description{
This function reorders columns and appends numeric suffixes. Don´t run it. It is an internal function.
}
\examples{
\dontrun{
library(dplyr)
# get_renamed0_df function reorders columns and appends numeric suffixes
}
}
\keyword{internal}
