% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAD_profiles.R
\docType{data}
\name{LAD_profiles}
\alias{LAD_profiles}
\title{LAD Profiles Dataset}
\format{
A data frame with columns: height, lad, treeID.
}
\usage{
LAD_profiles
}
\description{
Description of the LAD Profiles dataset.
}
\examples{
data(LAD_profiles)

}
\keyword{datasets}
