% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_no_flayer_noconsec.R
\name{remove_no_flayer_noconsec}
\alias{remove_no_flayer_noconsec}
\title{Remove non-consecutive fuel layers with a Leaf Area Density (LAD) less than a specified threshold}
\usage{
remove_no_flayer_noconsec(df, threshold = 25)
}
\arguments{
\item{df}{Internal data frame derived from [get_layers_lad()] function.}

\item{threshold}{Numeric value for the minimum required LAD percentage in a fuel layer. The default threshold is 25.}
}
\value{
No return value. The function is called for side effects.
}
\description{
This function removes non-consecutive fuel layers with a Leaf Area Density (LAD) less than a specified threshold and
recalculates the distances and the depth of remaining fuel layers. Don't run it. It is an internal function.
}
\examples{
# remove_no_flayer_noconsec() removes fuel layers with a LAD percentage less than a threshold
}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
