% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depths_calculation.R
\name{get_depths}
\alias{get_depths}
\title{Fuels depth in meters}
\usage{
get_depths (LAD_profiles, distance_metrics)
}
\arguments{
\item{LAD_profiles}{original tree Leaf Area Index (LAD) profile (output of [lad.profile()] function in the \emph{leafR} package.
An object of the class text}

\item{distance_metrics}{tree metrics with gaps (distances) and fuel base heights (output of [get_distance()] function).
An object of the class text}
}
\value{
A data frame giving fuel layers depth and the height of the depths in meters.
}
\description{
This function calculates fuels depth as the difference between gaps interleaved between fuel layers minus 1 if the fuel depths are greater than 1.
}
\details{
# List of tree metrics:
\itemize{
\item treeID: tree ID with strings and numeric values
\item treeID1: tree ID with only numeric values
\item cbh - Height of the fuel layers base height (m)
\item gap - Height of gaps between consecutive fuel layers (m)
\item dist: Distance between consecutive fuel layers (m)
\item Hdist - Height of the distance between consecutive fuel layers (m)
\item depth - Depth of fuel layers (m)
\item Hdepth - Height of the depth of fuel layers (m)
\item max_height - Maximum height of the tree profile
}
}
\examples{
## Not run:
library(dplyr)
library(magrittr)

# LAD profiles derived from normalized ALS data after applying [lad.profile()] function
data_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "LAD_profiles.txt")
LAD_profiles <- read.table(data_path, sep = "\t", header = TRUE)
LAD_profiles$treeID <- factor(LAD_profiles$treeID)

# Tree metrics derived from get_distance() function
distance_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "2_distance_metrics.txt")
distance_metrics <- read.table(distance_path, sep = "\t", header = TRUE)
distance_metrics$treeID <- factor(distance_metrics$treeID)

metrics_depth_list <- list()

for (i in levels(LAD_profiles$treeID)){

  tree1 <- LAD_profiles |> dplyr::filter(treeID == i)
  tree2 <- distance_metrics |> dplyr::filter(treeID == i)

  # Get depths for each tree
  metrics_depth <- get_depths(tree1, tree2)
  metrics_depth_list[[i]] <- metrics_depth
}

# Combine the individual data frames
metrics_all_depth <- dplyr::bind_rows(metrics_depth_list)

depth_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "3_depth_metrics.txt")
write.table(metrics_all_depth, file = depth_path, sep = "\t", row.names = FALSE)
## End(Not run)
}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
