% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_distances.R
\name{get_effective_gap}
\alias{get_effective_gap}
\title{Distances (> 1m) and CBHs (maximum and last distance)}
\usage{
get_effective_gap(depth_metrics_corr)
}
\arguments{
\item{depth_metrics_corr}{Tree metrics with the recalculated depth values considering distances > 1 m (output of [get_real_depths()] function).
An object of the class data frame.}
}
\value{
A data frame giving the effective distances (> 1 m) between consecutive fuel layers, identifying the CBH of the fuel layer
at the maximum distance and at the last distance.
}
\description{
This function recalculates the distance between fuel layers after removing distances = 1 m and determines the CBH based on the fuel layers
with the highest and the last distance.
}
\details{
List of tree metrics:
\itemize{
  \item treeID: tree ID with strings and numeric values
  \item treeID1: tree ID with only numeric values
  \item dist: Distance between consecutive fuel layers (m)
  \item Hdist: Height of the distance between consecutive fuel layers (m)
  \item Hcbh: Height of the base of each fuel layer (m)
  \item effdist: Effective distance between consecutive fuel layers (m) (> 1 m)
  \item dptf: Depth of fuel layers (m) after removing distances equal to 1 m
  \item Hdptf: Height of the depth of fuel layers (m) after removing distances equal to 1 m
  \item max_Hcbh: Canopy base height of the segmented tree based on the maximum distance found in its profile
  \item last_Hcbh: Canopy base height of the segmented tree based on the last distance found in its profile
  \item max_height: Maximum height of the tree profile
  \item max_: Values of distance and fuel depth and their corresponding heights at the maximum distance found in the tree profile
  \item last_: Values of distance and fuel depth and their corresponding heights at the last distance found in its profile
}
}
\examples{
## Not run:
library(dplyr)
library(magrittr)
library(stringr)

# Tree metrics derived from get_real_depths() function
depth_corr_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "5_tree_depth_metrics_corr.txt")
depth_metrics_corr <- read.table(depth_corr_path, sep = "\t", header = TRUE)
depth_metrics_corr$treeID <- factor(depth_metrics_corr$treeID)

trees_name1 <- as.character(depth_metrics_corr$treeID)
trees_name2 <- factor(unique(trees_name1))

distance_metrics_corr <- lapply(levels(trees_name2), function(i) {
  # Filter data for each tree
  tree2 <- depth_metrics_corr |> dplyr::filter(treeID == i)
  # Get effective gap for each tree
  get_effective_gap(tree2)
})

# Combine the individual data frames
distances_corr_all <- dplyr::bind_rows(distance_metrics_corr)

# =======================================================================#
# REORDER COLUMNS:
# =======================================================================#
# Get original column names
original_column_names <- colnames(distances_corr_all)

# Specify prefixes
prefixes <- c("treeID", "Hcbh", "dptf", "Hdptf", "effdist", "dist", "Hdist", "max_Hcbh", "max_dptf", "max_Hdptf", "last_Hcbh",
              "last_dptf", "last_Hdptf", "max_height")

# Initialize vector to store new order
new_order <- c()

# Loop over prefixes
for (prefix in prefixes) {
  # Find column names matching the current prefix
  matching_columns <- grep(paste0("^", prefix), original_column_names, value = TRUE)

  # Extract numeric suffixes and order the columns based on these suffixes
  numeric_suffixes <- as.numeric(gsub(paste0("^", prefix), "", matching_columns))
  matching_columns <- matching_columns[order(numeric_suffixes)]

  # Append to new order
  new_order <- c(new_order, matching_columns)
}

# Reorder values
distances_corr_all <- distances_corr_all[, new_order]
distance_corr_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "6_tree_distances_metrics_corr.txt")
write.table(distances_corr_all, file = distance_corr_path, sep = "\t", row.names = FALSE)
## End(Not run)

}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
