% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maxlad_metrics.R
\name{get_plots_cbh_LAD}
\alias{get_plots_cbh_LAD}
\title{Plots of fuel layers with LAD percentage > 5 and the canopy base height (CBH) based on the maximum LAD percentage}
\usage{
get_plots_cbh_LAD(LAD_profiles, fuels_LAD_metrics)
}
\arguments{
\item{LAD_profiles}{Original tree Leaf Area Index (LAD) profile (output of [lad.profile()] function in the \emph{leafR} package).
An object of the class text.}

\item{fuels_LAD_metrics}{Tree metrics with gaps (distances), fuel base heights, and depths of fuel layers with LAD percentage greater than 5
(output of [get_layers_lad()] function).
An object of the class text.}
}
\value{
A plot drawing the CBH of the fuel layer with the maximum LAD percentage and other fuel layers with LAD percentage greater than 5.
}
\description{
This function plots the CBH of the fuel layer with the maximum LAD percentage and other fuel layers with LAD percentage greater than 5.
}
\examples{
## Not run:
library(ggplot2)

# LAD profiles derived from normalized ALS data after applying [lad.profile()] function
data_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "LAD_profiles.txt")
LAD_profiles <- read.table(data_path, sep = "\t", header = TRUE)
LAD_profiles$treeID <- factor(LAD_profiles$treeID)

# Tree metrics derived from get_layers_lad() function
LAD_gt5p_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "7_fuels_lad_gt5perc.txt")
fuels_LAD_metrics <- read.table(LAD_gt5p_path, sep = "\t", header = TRUE)
fuels_LAD_metrics$treeID <- factor(fuels_LAD_metrics$treeID)

trees_name1 <- as.character(fuels_LAD_metrics$treeID)
trees_name2 <- factor(unique(trees_name1))

# Generate plots for fuels LAD metrics
plots_trees_LAD <- get_plots_cbh_LAD(LAD_profiles, fuels_LAD_metrics)

# Save plots for each tree
for (name in names(plots_trees_LAD)) {
  plots <- plots_trees_LAD[[name]]

  if (!is.null(plots)) {
    print(paste("Saving plot for tree:", name))
    ggsave(file.path(system.file("extdata", package = "LadderFuelsR"),  paste0( name, "_LAD_5perc", ".tiff")), plot = plots)
  }}
## End(Not run)

}
\author{
Olga Viedma, Carlos Silva and JM Moreno
}
