% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_depth.R
\name{get_real_depths}
\alias{get_real_depths}
\title{Fuel layers depth after removing distances = 1 m}
\usage{
get_real_depths (fbh_metrics_corr)
}
\arguments{
\item{fbh_metrics_corr}{tree metrics with the recalculated base height of fuel layers after considering distances > 1 m  (output of [get_real_fbh()] function).
An object of the class text}
}
\value{
A data frame giving the fuel layers depth after removing distances = 1 m.
}
\description{
This function recalculates fuel layers depth after removing distances = 1 m
}
\details{
# List of tree metrics:
\itemize{
\item treeID: tree ID with strings and numeric values
\item treeID1: tree ID with only numeric values
\item dist: Distance between consecutive fuel layers (m)
\item Hdist - Height of the distance between consecutive fuel layers (m)
\item Hcbh - Height of the base of each fuel layer (m)
\item dptf - Depth of fuel layers (m) after removing distances equal 1 m
\item Hdptf - Height of the depth of fuel layers (m) after removing distances equal 1 m
\item max_height - Maximum height of the tree profile
}
}
\examples{
## Not run:
library(dplyr)
library(magrittr)
library(tidyr)
# Tree metrics derived from get_real_fbh() function
fbhcor_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "4_fbh_metrics_corr.txt")
fbh_metrics_corr <- read.table(fbhcor_path, sep = "\t", header = TRUE)
fbh_metrics_corr$treeID <- factor(fbh_metrics_corr$treeID)

trees_name1 <- as.character(fbh_metrics_corr$treeID)
trees_name2 <- factor(unique(trees_name1))

depth_metrics_corr <- lapply(levels(trees_name2), function(i) {
  # Filter data for each tree
  tree2 <- fbh_metrics_corr |> dplyr::filter(treeID == i)
  # Get real depths for each tree
  get_real_depths(tree2)
})

# Combine depth values for all trees
depth_corr_all <- dplyr::bind_rows(depth_metrics_corr)
depth_corr_path <- file.path(system.file("extdata", package = "LadderFuelsR"), "5_tree_depth_metrics_corr.txt")
write.table(depth_corr_all, file = depth_corr_path, sep = "\t", row.names = FALSE)
## End(Not run)

}
\seealso{
\code{\link{get_renamed0_df}}
}
\author{
Olga Viedma.
}
