% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed_df.R
\name{get_renamed_df}
\alias{get_renamed_df}
\title{Rename and reorder columns (II)}
\usage{
get_renamed_df (df)
}
\arguments{
\item{df}{internal data frame derived from [get_layers_lad()] function}
}
\description{
This function deals with concatenated column names, reorders columns and appends numeric suffixes.
Don´t run it. It is an internal function.
}
\examples{
\dontrun{
library(dplyr)
# get_renamed_df concatenates column names, reorders columns and appends numeric suffixes
}
}
